Level 3 - Add hashing for passwords
Instead encryption added hash function. Installed  module "md5" which provides corresponding algorithm.
The password after registration converts by hash function into a new hash-code and stored in database in that state.
When the user tries to login, his typed password also hashed and searched in database for similar hash-code.
This approach is available, because for the same sequence of characters, hashing algorithm generates the same hash code.
Note: MD5 algorithm is absolutely insecure and can't be used in real projects.