Level 4 - Add hashing and salting with "bcrypt"
Added module "bcrypt" instead of "md5", as it gives higher reliability for the data.
This algorithm also provides "salting". Chosen 10 rounds.
"Bcrypt" algorithm is applied on the password value during the registration of a new user before it'll be added to the database.
In the login step "bcrypt" applied again for the typed by user password and compared with stored hash codes in the DB.