Level 5 - Add cookies and sessions using Passport.js
Added modules "passport", "passport-local", "passport-local-mongoose", "express-session".

"Passport" is Express-compatible authentication middleware for Node.js.
"Passport-Local" is a Passport strategy for authenticating with a username and password (the Local strategy).
"Passport-Local Mongoose" is a Mongoose plugin that simplifies building username and password login with Passport.
"Express-Session" creates a session middleware and it's not designed for a production environment, only for debugging and developing.

Implemented Local strategy for authentication: with username and password.
Previous hashing module "bcrypt" deleted as "Passport-Local Mongoose" hash and salt password field by default.
What changes do:
-set up the app to use sessions;
-initialize "Passport" and set it up to managing those sessions;
-set up the User schema to use the "Passport-Local Mongoose" as a plugin;
-use "Passport-Local Mongoose" to:
   +create a Local login strategy,
   +hash and salt user passwords;
   +save new users to the MongoDB database;
   +set up serialization and deserialization of users in shorter way (needs for sessions to create cookies with information
about Users and to use it for their authentication later on);
   +register new users;
   +authenticate already registered users (send to the browser a cookie with information that this User is authenticated and
 able get access to the "secrets.ejs" page);
-use "Passport" to establish a login session and in case of success - authenticate user;
-create a "/secrets" route because now due to sessions our users will be able to come back to this page until they log out
or close the browser (in real life, also after some period of time)
-create a "/logout" route to terminate a login session and to redirect the User to the root route ("/").

The order of the configuration code is important.