Level 6 - add OAuth authorization with Google or Github account
Added modules "passport-google-oauth20", "passport-github", "mongoose-findorcreate".

"passport-google-oauth20" - Passport strategy for authenticating with Google using the OAuth 2.0 API.
"passport-github" - Passport strategy for authenticating with GitHub using the OAuth 2.0 API.
"mongoose-findorcreate" - Mongoose plugin adds a "findOrCreate" method to models (required only to simplify code).

Implemented Passport strategies for authenticating with Google and Github using the OAuth 2.0 API.
Main steps:
1. Create an Application on Google ("Google Developers Console") and Github ("Developer applications"). After that you will get a client ID and client secret.
It needs to provide the strategy. You will also need to configure a redirect URI and else settings.
2. Configure Strategy. The client ID and secret obtained when creating an application are supplied as options when creating the strategy. These values were
added to ".env" file (environment variables). The strategies also require a verify callback. In the project added  plugin "mongoose-findorcreate", which implements
method "findOrCreate". In comments added also valid code as an alternative which does the same using only "Passport" and "Mongoose" functionalities.
3. Authenticate Requests. Added handler methods for GET requests to routes "/auth/google" and "/auth/github" with "passport.authenticate" method to specify
strategies and to redirect users to the google and github sign in pages respectively. If they sign in to their accounts successfully we will redirect them to the specified
earlier routes "/auth/google/secrets" and "/auth/github/secrets" to authenticate them on our web site with received from Google/Github access tokens .
(This routes we specified when we were created new projects, for instance, in the "Google Developers Console" ("http://localhost:3000/auth/google/secrets" in that case))

Also added to the files "login.ejs" and "register.ejs" buttons "Sign In with Google" and "Sign In with Github".
To style them from the package "Social Buttons for Bootstrap" added file "bootstrap-social.css" and a link to it in the "header.ejs" file.